SignalR ile gerçek zamanlı tablo güncellemeleri ve CSV ile toplu veri yükleme eklendi
- Sunucu ile istemci arasında anlık çift yönlü iletişim kurmak için projeye SignalR (WeatherHub) entegre edildi.
- Q1Worker ve Q2Worker işçileri, API'den veri çektikleri an 'ReceiveWeatherResult' event'i ile sonuçları tarayıcıya fırlatacak şekilde güncellendi.
- index.html arayüzüne CSV dosya seçici eklendi ve '/api/upload-csv' endpointi oluşturuldu (Minimal API Antiforgery koruması bypass edildi).
- Farklı hava durumu servislerinin aynı lokasyon için farklı şehir isimleri döndürmesinden kaynaklanan HTML satır duplikasyonu hatası düzeltildi; tablo satır ID'leri (DOM Elements) isim yerine Enlem/Boylam koordinatlarına bağlandı.

+++ WeatherCollector/DatabaseScripts/03-StoredProcedure.sql
@@ -22,8 +22,8 @@ GO
 
 CREATE PROCEDURE dbo.SaveResults
 	@Name NVARCHAR(255),
-    @Latitude DECIMAL(10, 6),
-	@Longitude DECIMAL(10, 6),
+    @Latitude Float(10, 6),
+	@Longitude Float(10, 6),
 	@Temperature FLOAT,
 	@ThreadName NVARCHAR(50)
 AS
